+++ webviewsrc/util/vscode.ts
@@ -1,2 +1,7 @@
-declare const acquireVsCodeApi: any;
+interface VSCodeApi {
+    postMessage<T>(message: T): void;
+    getState(): any;
+    setState(state: any): void;
+}
+declare function acquireVsCodeApi(): VSCodeApi;
 export const vscode = acquireVsCodeApi();